requirements.txt
biopython==1.81
numpy==1.26.0
